+++ src/main/resources/templates/login.html
@@ -41,7 +41,7 @@
                                                             <h2 class="text-center no-mrg-vertical mrg-btm-15 text-dark">
                                                                 로그인</h2>
                                                         </div>
-                                                        <form method="post" action="/login/check">
+                                                        <form method="post" action="/login">
                                                             <div class="form-group">
                                                                 <input class="form-control" placeholder="Id"
                                                                        name="email"
