+++ src/test/java/techcourse/myblog/web/LoginControllerTest.java
@@ -7,7 +7,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.web.reactive.function.BodyInserters;
-import techcourse.myblog.user.User;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureWebTestClient
@@ -15,32 +14,6 @@ public class LoginControllerTest {
 
     @Autowired
     WebTestClient webTestClient;
-
-    void setSession() {
-        webTestClient = WebTestClient.bindToWebHandler(exchange -> {
-            String path = exchange.getRequest().getURI().getPath();
-            if ("/users/mypage".equals(path) || "/users/mypage/edit".equals(path)) {
-                return exchange.getSession()
-                        .doOnNext(webSession ->
-                                webSession.getAttributes().put("user", new User("buddy", "buddyCU@buddy.com", "Aa12345!")))
-                        .then();
-            }
-            return null;
-        }).build();
-    }
-
-    void removeSession() {
-        webTestClient = WebTestClient.bindToWebHandler(exchange -> {
-            String path = exchange.getRequest().getURI().getPath();
-            if ("/users/mypage".equals(path) || "/users/mypage/edit".equals(path)) {
-                return exchange.getSession()
-                        .doOnNext(webSession ->
-                                webSession.getAttributes().remove("user"))
-                        .then();
-            }
-            return null;
-        }).build();
-    }
 
     @Test
     void 로그인_페이지_접근() {
@@ -66,7 +39,7 @@ public class LoginControllerTest {
                         .with("confirmPassword", "Aa12345!")
                 ).exchange();
 
-        webTestClient.post().uri("/login/check")
+        webTestClient.post().uri("/login")
                 .contentType(MediaType.APPLICATION_FORM_URLENCODED)
                 .body(BodyInserters
                         .fromFormData("email", "buddy@buddy.com")
@@ -86,7 +59,7 @@ public class LoginControllerTest {
                         .with("confirmPassword", "Aa12345!")
                 ).exchange();
 
-        webTestClient.post().uri("/login/check")
+        webTestClient.post().uri("/login")
                 .contentType(MediaType.APPLICATION_FORM_URLENCODED)
                 .body(BodyInserters
                         .fromFormData("email", "ssosso@gmail.com")
@@ -94,34 +67,6 @@ public class LoginControllerTest {
                 .exchange()
                 .expectStatus()
                 .isBadRequest();
-    }
-
-    @Test
-    void 로그인_전_마이페이지_접근() {
-        webTestClient.get().uri("/users/mypage").exchange()
-                .expectStatus().isBadRequest();
-    }
-
-    @Test
-    void 로그인_후_마이페이지_접근() {
-        setSession();
-        webTestClient.get().uri("/users/mypage").exchange()
-                .expectStatus().isOk();
-        removeSession();
-    }
-
-    @Test
-    void 로그인_전_회원수정페이지_접근() {
-        webTestClient.get().uri("/users/mypage/edit").exchange()
-                .expectStatus().isBadRequest();
-    }
-
-    @Test
-    void 로그인_후_회원수정페이지_접근() {
-        setSession();
-        webTestClient.get().uri("/users/mypage/edit").exchange()
-                .expectStatus().isOk();
-        removeSession();
     }
 
 }
